+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 3.5.1] 2025-07-09
+
+### Fixed
+
+- ETL_engineering: Fixed a bug in `KeywordDummifier` trying to use DataFrame constructor with a polars expression.
+- data_exploration: Fixed a bug in `drop_zero_only_columns()` that prevented it to handle NaN values.
+
 ## [3.5.0] 2025-07-09
 
 ### Changed